Remove <time> tags
At least one of the timestamps in this article is outside the bounds of
what browsers can handle in a <time> tag's datetime attribute; however,
<time> elements without a datetime attribute must have no descendent
elements (e.g., using <abbr> inside them is illegal).

This is understandable, but makes it impossible to use the formatting
that I would like, so no <time> tags for us.

Signed-off-by: Sam Stuewe <samuel.stuewe@gmail.com>